requirements.txt
Django
Celery
beautifulsoup4
djangorestframework
requests